sync 19893: fix Wformat warnings

+++ tests/cpp-tests/Classes/FileUtilsTest/FileUtilsTest.cpp
@@ -1322,7 +1322,7 @@ void TestListFiles::onEnter()
     std::vector<std::string> list = FileUtils::getInstance()->listFiles (defaultPath);
 
     char cntBuffer[200] = { 0 };
-    snprintf(cntBuffer, 200, "'fonts/' %d, $defaultResourceRootPath %d", (int)listFonts.size(), (int)list.size());
+    snprintf(cntBuffer, 200, "'fonts/' %zu, $defaultResourceRootPath %zu",listFonts.size(), list.size());
 
     for(int i=0;i<listFonts.size();i++)
     {
